Move `RedisServer` to redis-test.
