+++ redis-test/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "redis-test"
-version = "0.6.0"
+version = "0.7.0"
 edition = "2021"
 description = "Testing helpers for the `redis` crate"
 homepage = "https://github.com/redis-rs/redis-rs"
@@ -14,13 +14,23 @@ bench = false
 
 [dependencies]
 redis = { version = "0.27.0", path = "../redis" }
-
 bytes = { version = "1", optional = true }
 futures = { version = "0.3", optional = true }
+tempfile = "=3.14.0"
+socket2 = "0.5"
+rand = "0.8.5"
 
 [features]
 aio = ["futures", "redis/aio"]
 
 [dev-dependencies]
-redis = { version = "0.27.0", path = "../redis", features = ["aio", "tokio-comp"] }
-tokio = { version = "1", features = ["rt", "macros", "rt-multi-thread", "time"] }
+redis = { version = "0.27.0", path = "../redis", features = [
+    "aio",
+    "tokio-comp",
+] }
+tokio = { version = "1", features = [
+    "rt",
+    "macros",
+    "rt-multi-thread",
+    "time",
+] }
