+++ redis-test/src/lib.rs
@@ -23,6 +23,9 @@
 //! assert_eq!(result, true);
 //! ```
 
+pub mod server;
+pub mod utils;
+
 use std::collections::VecDeque;
 use std::sync::{Arc, Mutex};
