+++ redis/tests/support/mod.rs
@@ -1,37 +1,21 @@
 #![allow(dead_code)]
 
-use std::path::Path;
-use std::{
-    env, fs, io, net::SocketAddr, net::TcpListener, path::PathBuf, process, thread::sleep,
-    time::Duration,
-};
+#[cfg(feature = "aio")]
+use futures::Future;
+#[cfg(feature = "aio")]
+use redis::{aio, cmd};
+use redis::{ConnectionAddr, InfoDict, Pipeline, ProtocolVersion, RedisResult, Value};
+use redis_test::server::{use_protocol, Module, RedisServer};
+use redis_test::utils::{get_random_available_port, TlsFilePaths};
 #[cfg(feature = "tls-rustls")]
 use std::{
     fs::File,
     io::{BufReader, Read},
 };
-
-#[cfg(feature = "aio")]
-use futures::Future;
-#[cfg(feature = "aio")]
-use redis::{aio, cmd};
-use redis::{
-    ConnectionAddr, InfoDict, Pipeline, ProtocolVersion, RedisConnectionInfo, RedisResult, Value,
-};
+use std::{io, thread::sleep, time::Duration};
 
 #[cfg(feature = "tls-rustls")]
 use redis::{ClientTlsConfig, TlsCertificates};
-
-use socket2::{Domain, Socket, Type};
-use tempfile::TempDir;
-
-pub fn use_protocol() -> ProtocolVersion {
-    if env::var("PROTOCOL").unwrap_or_default() == "RESP3" {
-        ProtocolVersion::RESP3
-    } else {
-        ProtocolVersion::RESP2
-    }
-}
 
 pub fn current_thread_runtime() -> tokio::runtime::Runtime {
     let mut builder = tokio::runtime::Builder::new_current_thread();
@@ -149,291 +133,6 @@ mod sentinel;
 #[allow(unused_imports)]
 pub use self::sentinel::*;
 
-#[derive(PartialEq)]
-enum ServerType {
-    Tcp { tls: bool },
-    Unix,
-}
-
-pub enum Module {
-    Json,
-}
-
-pub struct RedisServer {
-    pub process: process::Child,
-    tempdir: tempfile::TempDir,
-    log_file: PathBuf,
-    addr: redis::ConnectionAddr,
-    pub(crate) tls_paths: Option<TlsFilePaths>,
-}
-
-impl ServerType {
-    fn get_intended() -> ServerType {
-        match env::var("REDISRS_SERVER_TYPE")
-            .ok()
-            .as_ref()
-            .map(|x| &x[..])
-        {
-            Some("tcp") => ServerType::Tcp { tls: false },
-            Some("tcp+tls") => ServerType::Tcp { tls: true },
-            Some("unix") => ServerType::Unix,
-            Some(val) => {
-                panic!("Unknown server type {val:?}");
-            }
-            None => ServerType::Tcp { tls: false },
-        }
-    }
-}
-
-impl RedisServer {
-    pub fn new() -> RedisServer {
-        RedisServer::with_modules(&[], false)
-    }
-
-    #[cfg(feature = "tls-rustls")]
-    pub fn new_with_mtls() -> RedisServer {
-        RedisServer::with_modules(&[], true)
-    }
-
-    pub fn log_file_contents(&self) -> Option<String> {
-        std::fs::read_to_string(self.log_file.clone()).ok()
-    }
-
-    pub fn get_addr(port: u16) -> ConnectionAddr {
-        let server_type = ServerType::get_intended();
-        match server_type {
-            ServerType::Tcp { tls } => {
-                if tls {
-                    redis::ConnectionAddr::TcpTls {
-                        host: "127.0.0.1".to_string(),
-                        port,
-                        insecure: true,
-                        tls_params: None,
-                    }
-                } else {
-                    redis::ConnectionAddr::Tcp("127.0.0.1".to_string(), port)
-                }
-            }
-            ServerType::Unix => {
-                let (a, b) = rand::random::<(u64, u64)>();
-                let path = format!("/tmp/redis-rs-test-{a}-{b}.sock");
-                redis::ConnectionAddr::Unix(PathBuf::from(&path))
-            }
-        }
-    }
-
-    pub fn with_modules(modules: &[Module], mtls_enabled: bool) -> RedisServer {
-        // this is technically a race but we can't do better with
-        // the tools that redis gives us :(
-        let redis_port = get_random_available_port();
-        let addr = RedisServer::get_addr(redis_port);
-
-        RedisServer::new_with_addr_tls_modules_and_spawner(
-            addr,
-            None,
-            None,
-            mtls_enabled,
-            modules,
-            |cmd| {
-                cmd.spawn()
-                    .unwrap_or_else(|err| panic!("Failed to run {cmd:?}: {err}"))
-            },
-        )
-    }
-
-    pub fn new_with_addr_and_modules(
-        addr: redis::ConnectionAddr,
-        modules: &[Module],
-        mtls_enabled: bool,
-    ) -> RedisServer {
-        RedisServer::new_with_addr_tls_modules_and_spawner(
-            addr,
-            None,
-            None,
-            mtls_enabled,
-            modules,
-            |cmd| {
-                cmd.spawn()
-                    .unwrap_or_else(|err| panic!("Failed to run {cmd:?}: {err}"))
-            },
-        )
-    }
-
-    pub fn new_with_addr_tls_modules_and_spawner<
-        F: FnOnce(&mut process::Command) -> process::Child,
-    >(
-        addr: redis::ConnectionAddr,
-        config_file: Option<&Path>,
-        tls_paths: Option<TlsFilePaths>,
-        mtls_enabled: bool,
-        modules: &[Module],
-        spawner: F,
-    ) -> RedisServer {
-        #[cfg(feature = "tls-rustls")]
-        if rustls::crypto::CryptoProvider::get_default().is_none() {
-            // we don't care about success, because failure means that the provider was set from another thread.
-            let _ = rustls::crypto::ring::default_provider().install_default();
-        }
-
-        let mut redis_cmd = process::Command::new("redis-server");
-
-        if let Some(config_path) = config_file {
-            redis_cmd.arg(config_path);
-        }
-
-        // Load Redis Modules
-        for module in modules {
-            match module {
-                Module::Json => {
-                    redis_cmd
-                        .arg("--loadmodule")
-                        .arg(env::var("REDIS_RS_REDIS_JSON_PATH").expect(
-                        "Unable to find path to RedisJSON at REDIS_RS_REDIS_JSON_PATH, is it set?",
-                    ));
-                }
-            };
-        }
-
-        redis_cmd
-            .stdout(process::Stdio::piped())
-            .stderr(process::Stdio::piped());
-        let tempdir = tempfile::Builder::new()
-            .prefix("redis")
-            .tempdir()
-            .expect("failed to create tempdir");
-        let log_file = Self::log_file(&tempdir);
-        redis_cmd.arg("--logfile").arg(log_file.clone());
-        match addr {
-            redis::ConnectionAddr::Tcp(ref bind, server_port) => {
-                redis_cmd
-                    .arg("--port")
-                    .arg(server_port.to_string())
-                    .arg("--bind")
-                    .arg(bind);
-
-                RedisServer {
-                    process: spawner(&mut redis_cmd),
-                    log_file,
-                    tempdir,
-                    addr,
-                    tls_paths: None,
-                }
-            }
-            redis::ConnectionAddr::TcpTls { ref host, port, .. } => {
-                let tls_paths = tls_paths.unwrap_or_else(|| build_keys_and_certs_for_tls(&tempdir));
-
-                let auth_client = if mtls_enabled { "yes" } else { "no" };
-
-                // prepare redis with TLS
-                redis_cmd
-                    .arg("--tls-port")
-                    .arg(port.to_string())
-                    .arg("--port")
-                    .arg("0")
-                    .arg("--tls-cert-file")
-                    .arg(&tls_paths.redis_crt)
-                    .arg("--tls-key-file")
-                    .arg(&tls_paths.redis_key)
-                    .arg("--tls-ca-cert-file")
-                    .arg(&tls_paths.ca_crt)
-                    .arg("--tls-auth-clients")
-                    .arg(auth_client)
-                    .arg("--bind")
-                    .arg(host);
-
-                // Insecure only disabled if `mtls` is enabled
-                let insecure = !mtls_enabled;
-
-                let addr = redis::ConnectionAddr::TcpTls {
-                    host: host.clone(),
-                    port,
-                    insecure,
-                    tls_params: None,
-                };
-
-                RedisServer {
-                    process: spawner(&mut redis_cmd),
-                    log_file,
-                    tempdir,
-                    addr,
-                    tls_paths: Some(tls_paths),
-                }
-            }
-            redis::ConnectionAddr::Unix(ref path) => {
-                redis_cmd
-                    .arg("--port")
-                    .arg("0")
-                    .arg("--unixsocket")
-                    .arg(path);
-                RedisServer {
-                    process: spawner(&mut redis_cmd),
-                    log_file,
-                    tempdir,
-                    addr,
-                    tls_paths: None,
-                }
-            }
-        }
-    }
-
-    pub fn client_addr(&self) -> &redis::ConnectionAddr {
-        &self.addr
-    }
-
-    pub fn connection_info(&self) -> redis::ConnectionInfo {
-        redis::ConnectionInfo {
-            addr: self.client_addr().clone(),
-            redis: RedisConnectionInfo {
-                protocol: use_protocol(),
-                ..Default::default()
-            },
-        }
-    }
-
-    pub fn stop(&mut self) {
-        let _ = self.process.kill();
-        let _ = self.process.wait();
-        if let redis::ConnectionAddr::Unix(ref path) = *self.client_addr() {
-            fs::remove_file(path).ok();
-        }
-    }
-
-    pub fn log_file(tempdir: &TempDir) -> PathBuf {
-        tempdir.path().join("redis.log")
-    }
-}
-
-pub fn get_listener_on_free_port() -> TcpListener {
-    let addr = &"127.0.0.1:0".parse::<SocketAddr>().unwrap().into();
-    let socket = Socket::new(Domain::IPV4, Type::STREAM, None).unwrap();
-    socket.set_reuse_address(true).unwrap();
-    socket.bind(addr).unwrap();
-    socket.listen(1).unwrap();
-    TcpListener::from(socket)
-}
-
-/// Finds a random open port available for listening at, by spawning a TCP server with
-/// port "zero" (which prompts the OS to just use any available port). Between calling
-/// this function and trying to bind to this port, the port may be given to another
-/// process, so this must be used with care (since here we only use it for tests, it's
-/// mostly okay).
-pub fn get_random_available_port() -> u16 {
-    for _ in 0..10000 {
-        let listener = get_listener_on_free_port();
-        let port = listener.local_addr().unwrap().port();
-        if port < 55535 {
-            return port;
-        }
-    }
-    panic!("Couldn't get a valid port");
-}
-
-impl Drop for RedisServer {
-    fn drop(&mut self) {
-        self.stop()
-    }
-}
-
 pub struct TestContext {
     pub server: RedisServer,
     pub client: redis::Client,
@@ -442,6 +141,14 @@ pub struct TestContext {
 
 pub(crate) fn is_tls_enabled() -> bool {
     cfg!(all(feature = "tls-rustls", not(feature = "tls-native-tls")))
+}
+
+pub(crate) fn start_tls_crypto_provider() {
+    #[cfg(feature = "tls-rustls")]
+    if rustls::crypto::CryptoProvider::get_default().is_none() {
+        // we don't care about success, because failure means that the provider was set from another thread.
+        let _ = rustls::crypto::ring::default_provider().install_default();
+    }
 }
 
 impl TestContext {
@@ -471,6 +178,7 @@ impl TestContext {
         mtls_enabled: bool,
         tls_files: Option<TlsFilePaths>,
     ) -> Self {
+        start_tls_crypto_provider();
         let redis_port = get_random_available_port();
         let addr = RedisServer::get_addr(redis_port);
         Self::with_modules_addr_and_tls(modules, mtls_enabled, addr, tls_files)
@@ -646,124 +354,6 @@ where
             Some(details) => write!(writer, "-{} {details}\r\n", err.code()),
             None => write!(writer, "-{}\r\n", err.code()),
         },
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct TlsFilePaths {
-    pub(crate) redis_crt: PathBuf,
-    pub(crate) redis_key: PathBuf,
-    pub(crate) ca_crt: PathBuf,
-}
-
-pub fn build_keys_and_certs_for_tls(tempdir: &TempDir) -> TlsFilePaths {
-    // Based on shell script in redis's server tests
-    // https://github.com/redis/redis/blob/8c291b97b95f2e011977b522acf77ead23e26f55/utils/gen-test-certs.sh
-    let ca_crt = tempdir.path().join("ca.crt");
-    let ca_key = tempdir.path().join("ca.key");
-    let ca_serial = tempdir.path().join("ca.txt");
-    let redis_crt = tempdir.path().join("redis.crt");
-    let redis_key = tempdir.path().join("redis.key");
-    let ext_file = tempdir.path().join("openssl.cnf");
-
-    fn make_key<S: AsRef<std::ffi::OsStr>>(name: S, size: usize) {
-        process::Command::new("openssl")
-            .arg("genrsa")
-            .arg("-out")
-            .arg(name)
-            .arg(format!("{size}"))
-            .stdout(process::Stdio::piped())
-            .stderr(process::Stdio::piped())
-            .spawn()
-            .expect("failed to spawn openssl")
-            .wait()
-            .expect("failed to create key");
-    }
-
-    // Build CA Key
-    make_key(&ca_key, 4096);
-
-    // Build redis key
-    make_key(&redis_key, 2048);
-
-    // Build CA Cert
-    process::Command::new("openssl")
-        .arg("req")
-        .arg("-x509")
-        .arg("-new")
-        .arg("-nodes")
-        .arg("-sha256")
-        .arg("-key")
-        .arg(&ca_key)
-        .arg("-days")
-        .arg("3650")
-        .arg("-subj")
-        .arg("/O=Redis Test/CN=Certificate Authority")
-        .arg("-out")
-        .arg(&ca_crt)
-        .stdout(process::Stdio::piped())
-        .stderr(process::Stdio::piped())
-        .spawn()
-        .expect("failed to spawn openssl")
-        .wait()
-        .expect("failed to create CA cert");
-
-    // Build x509v3 extensions file
-    fs::write(
-        &ext_file,
-        b"keyUsage = digitalSignature, keyEncipherment\n\
-    subjectAltName = @alt_names\n\
-    [alt_names]\n\
-    IP.1 = 127.0.0.1\n",
-    )
-    .expect("failed to create x509v3 extensions file");
-
-    // Read redis key
-    let mut key_cmd = process::Command::new("openssl")
-        .arg("req")
-        .arg("-new")
-        .arg("-sha256")
-        .arg("-subj")
-        .arg("/O=Redis Test/CN=Generic-cert")
-        .arg("-key")
-        .arg(&redis_key)
-        .stdout(process::Stdio::piped())
-        .stderr(process::Stdio::piped())
-        .spawn()
-        .expect("failed to spawn openssl");
-
-    // build redis cert
-    process::Command::new("openssl")
-        .arg("x509")
-        .arg("-req")
-        .arg("-sha256")
-        .arg("-CA")
-        .arg(&ca_crt)
-        .arg("-CAkey")
-        .arg(&ca_key)
-        .arg("-CAserial")
-        .arg(&ca_serial)
-        .arg("-CAcreateserial")
-        .arg("-days")
-        .arg("365")
-        .arg("-extfile")
-        .arg(&ext_file)
-        .arg("-out")
-        .arg(&redis_crt)
-        .stdin(key_cmd.stdout.take().expect("should have stdout"))
-        .stdout(process::Stdio::piped())
-        .stderr(process::Stdio::piped())
-        .spawn()
-        .expect("failed to spawn openssl")
-        .wait()
-        .expect("failed to create redis cert");
-
-    key_cmd.wait().expect("failed to create redis key");
-
-    TlsFilePaths {
-        redis_crt,
-        redis_key,
-        ca_crt,
     }
 }
