+++ redis/tests/test_cluster.rs
@@ -15,6 +15,7 @@ mod cluster {
         cmd, parse_redis_value, Commands, ConnectionLike, ErrorKind, ProtocolVersion, RedisError,
         Value,
     };
+    use redis_test::server::use_protocol;
 
     #[test]
     fn test_cluster_basics() {
