+++ redis/tests/test_cluster_async.rs
@@ -25,6 +25,7 @@ mod cluster_async {
         InfoDict, IntoConnectionInfo, ProtocolVersion, RedisError, RedisFuture, RedisResult,
         Script, Value,
     };
+    use redis_test::server::use_protocol;
     use rstest::rstest;
 
     use crate::support::*;
